Graphics.Gloss.Juicy

Conversion from JuicyPixels' types to gloss' Picture

fromDynamicImage :: DynamicImage -> Maybe Picture

Tries to convert a DynamicImage from JuicyPixels to a gloss Picture. All formats except RGBF and YF should successfully yield a Picture.

fromImageRGBA8 :: Image PixelRGBA8 -> Picture

O(N) conversion from PixelRGBA8 image to gloss Picture, where N is the number of pixels.

fromImageRGB8 :: Image PixelRGB8 -> Picture

Creation of a gloss Picture by promoting (through promoteImage) the PixelRGB8 image to PixelRGBA8 and calling fromImageRGBA8.

fromImageY8 :: Image Pixel8 -> Picture

Creation of a gloss Picture by promoting (through promoteImage) the PixelY8 image to PixelRGBA8 and calling fromImageRGBA8.

fromImageYA8 :: Image PixelYA8 -> Picture

Creation of a gloss Picture by promoting (through promoteImage) the PixelYA8 image to PixelRGBA8 and calling fromImageRGBA8.

fromImageYCbCr8 :: Image PixelYCbCr8 -> Picture

Creation of a gloss Picture by promoting (through promoteImage) the PixelYCbCr8 image to PixelRGBA8 and calling fromImageRGBA8.

Loading a gloss Picture from a file through JuicyPixels

loadJuicy :: FilePath -> IO (Maybe Picture)

Tries to load an image file into a Picture using readImage from JuicyPixels. It means it'll try to successively read the content as an image in the following order, until it succeeds (or fails on all of them): jpeg, png, bmp, gif, hdr (the last two are not supported) This is handy when you don't know what format the image contained in the file is encoded with. If you know the format in advance, use loadBMP, loadJuicyJPG or loadJuicyPNG

loadJuicyJPG :: FilePath -> IO (Maybe Picture)

loadJuicyPNG :: FilePath -> IO (Maybe Picture)

From gloss, exported here for convenience

loadBMP :: FilePath -> IO Picture

Load an uncompressed 24 or 32bit RGBA BMP file as a bitmap.